moved PurchasedProduct to common module
