+++ common/src/main/java/com/revenuecat/purchases/common/offlineentitlements/CustomerInfoResponseHandler.kt
@@ -10,7 +10,6 @@ import com.revenuecat.purchases.common.CustomerInfoFactory
 import com.revenuecat.purchases.common.DateProvider
 import com.revenuecat.purchases.common.DefaultDateProvider
 import com.revenuecat.purchases.common.errorLog
-import com.revenuecat.purchases.models.PurchasedProduct
 import com.revenuecat.purchases.strings.CustomerInfoStrings.COMPUTING_OFFLINE_CUSTOMER_INFO_FAILED
 import com.revenuecat.purchases.utils.Iso8601Utils
 import org.json.JSONObject
