+++ common/src/test/java/com/revenuecat/purchases/common/offlineentitlements/CustomerInfoResponseHandlerTest.kt
@@ -8,9 +8,7 @@ import com.revenuecat.purchases.PeriodType
 import com.revenuecat.purchases.Store
 import com.revenuecat.purchases.common.BillingAbstract
 import com.revenuecat.purchases.common.DateProvider
-import com.revenuecat.purchases.models.PurchasedProduct
 import com.revenuecat.purchases.utils.dummyGoogleRestoredStoreTransaction
-import com.revenuecat.purchases.utils.stubGooglePurchase
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.Before
