+++ feature/amazon/src/main/java/com/revenuecat/purchases/amazon/AmazonBilling.kt
@@ -33,7 +33,7 @@ import com.revenuecat.purchases.common.errorLog
 import com.revenuecat.purchases.common.log
 import com.revenuecat.purchases.common.sha1
 import com.revenuecat.purchases.models.PurchaseState
-import com.revenuecat.purchases.models.PurchasedProduct
+import com.revenuecat.purchases.common.offlineentitlements.PurchasedProduct
 import com.revenuecat.purchases.models.StoreProduct
 import com.revenuecat.purchases.models.StoreTransaction
 import com.revenuecat.purchases.strings.PurchaseStrings
